requirements.txt
# dev environment
setuptools~=60.2
twine~=4.0
sphinx~=5.3
sphinx-rtd-theme~=1.1
# Jupyter notebooks
pytz~=2022.2
# true package dependencies
requests~=2.28
pandas~=1.3
numpy~=1.21 ; python_version < '3.8'
numpy~=1.22 ; python_version >= '3.8'
boto3~=1.24
botocore~=1.27
google-cloud-storage~=2.7
azure-storage-blob~=12.14
azure-identity~=1.12
pyodk~=0.1
python-dateutil~=2.8.2
flatten_json~=0.1